requirements.txt
streamlit
psycopg2-binary
pandas
bcrypt
python-dotenv
matplotlib
